+++ CoffeeBeanExplorer.Infrastructure/Repositories/UserRepository.cs
@@ -30,8 +30,7 @@ public class UserRepository(DbConnectionFactory dbContext) : IUserRepository
             """,
             new { Id = id });
 
-        if (user != null)
-        {
+        if (user is not null)
             user.RefreshTokens = (await connection.QueryAsync<RefreshToken>(
                 """
                 SELECT "Id", "Token", "Expires", "Created", "Revoked", "ReplacedByToken", "ReasonRevoked", "UserId"
@@ -39,7 +38,6 @@ public class UserRepository(DbConnectionFactory dbContext) : IUserRepository
                 WHERE "UserId" = @UserId
                 """,
                 new { UserId = id })).ToList();
-        }
 
         return user;
     }
@@ -56,8 +54,7 @@ public class UserRepository(DbConnectionFactory dbContext) : IUserRepository
             """,
             new { Username = username });
 
-        if (user != null)
-        {
+        if (user is not null)
             user.RefreshTokens = (await connection.QueryAsync<RefreshToken>(
                 """
                 SELECT "Id", "Token", "Expires", "Created", "Revoked", "ReplacedByToken", "ReasonRevoked", "UserId"
@@ -65,7 +62,6 @@ public class UserRepository(DbConnectionFactory dbContext) : IUserRepository
                 WHERE "UserId" = @UserId
                 """,
                 new { UserId = user.Id })).ToList();
-        }
 
         return user;
     }
@@ -82,8 +78,7 @@ public class UserRepository(DbConnectionFactory dbContext) : IUserRepository
             """,
             new { Email = email });
 
-        if (user != null)
-        {
+        if (user is not null)
             user.RefreshTokens = (await connection.QueryAsync<RefreshToken>(
                 """
                 SELECT "Id", "Token", "Expires", "Created", "Revoked", "ReplacedByToken", "ReasonRevoked", "UserId"
@@ -91,7 +86,6 @@ public class UserRepository(DbConnectionFactory dbContext) : IUserRepository
                 WHERE "UserId" = @UserId
                 """,
                 new { UserId = user.Id })).ToList();
-        }
 
         return user;
     }
@@ -115,13 +109,23 @@ public class UserRepository(DbConnectionFactory dbContext) : IUserRepository
         using var connection = dbContext.GetConnection();
         var insertedUserId = await connection.QuerySingleAsync<int>(
             """
-            INSERT INTO "Auth"."Users" ("Username", "Email", "FirstName", "LastName", "PasswordHash", 
-                                        "Salt", "IsActive", "CreatedAt", "UpdatedAt")
-            VALUES (@Username, @Email, @FirstName, @LastName, @PasswordHash, 
-                    @Salt, @IsActive, now(), now())
+            INSERT INTO "Auth"."Users" ("Username", "Email", "FirstName", "LastName", "PasswordHash",
+                                        "Salt", "IsActive", "CreatedAt", "UpdatedAt", "Role")
+            VALUES (@Username, @Email, @FirstName, @LastName, @PasswordHash,
+                    @Salt, @IsActive, now(), now(), @Role::"UserRole")
             RETURNING "Id"
             """,
-            user);
+            new
+            {
+                user.Username,
+                user.Email,
+                user.FirstName,
+                user.LastName,
+                user.PasswordHash,
+                user.Salt,
+                user.IsActive,
+                Role = user.Role.ToString()
+            });
 
         user.Id = insertedUserId;
 
@@ -163,11 +167,8 @@ public class UserRepository(DbConnectionFactory dbContext) : IUserRepository
                 transaction);
 
             if (user.RefreshTokens.Count != 0)
-            {
                 foreach (var token in user.RefreshTokens)
-                {
                     if (token.Id == 0)
-                    {
                         await connection.ExecuteAsync(
                             """
                             INSERT INTO "Auth"."RefreshTokens" ("Token", "Expires", "Created", "Revoked", "ReplacedByToken", "ReasonRevoked", "UserId")
@@ -175,9 +176,7 @@ public class UserRepository(DbConnectionFactory dbContext) : IUserRepository
                             """,
                             token,
                             transaction);
-                    }
                     else
-                    {
                         await connection.ExecuteAsync(
                             """
                             UPDATE "Auth"."RefreshTokens"
@@ -188,9 +187,6 @@ public class UserRepository(DbConnectionFactory dbContext) : IUserRepository
                             """,
                             token,
                             transaction);
-                    }
-                }
-            }
 
             transaction.Commit();
             return rowsAffected > 0;
@@ -205,31 +201,44 @@ public class UserRepository(DbConnectionFactory dbContext) : IUserRepository
     public async Task<bool> DeleteAsync(int id)
     {
         using var connection = dbContext.GetConnection();
-        using var transaction = connection.BeginTransaction();
+        var userExists = await connection.ExecuteScalarAsync<bool>(
+            "SELECT COUNT(*) > 0 FROM \"Auth\".\"Users\" WHERE \"Id\" = @Id",
+            new { Id = id });
 
+        if (!userExists) return false;
+
+        using var transaction = connection.BeginTransaction();
         try
         {
             await connection.ExecuteAsync(
                 """
+                DELETE FROM "Social"."ListItems" WHERE "ListId" IN (SELECT "Id" FROM "Social"."UserLists" WHERE "UserId" = @Id);
                 DELETE FROM "Auth"."RefreshTokens" WHERE "UserId" = @Id;
                 DELETE FROM "Social"."UserLists" WHERE "UserId" = @Id;
                 DELETE FROM "Social"."Reviews" WHERE "UserId" = @Id;
                 DELETE FROM "Auth"."Users" WHERE "Id" = @Id;
                 """,
-                new { Id = id },
-                transaction);
+                new { Id = id }, transaction);
 
             transaction.Commit();
-            var exists = await connection.ExecuteScalarAsync<bool>(
-                "SELECT COUNT(*) > 0 FROM \"Auth\".\"Users\" WHERE \"Id\" = @Id",
-                new { Id = id });
-
-            return !exists;
+            return true;
         }
         catch
         {
             transaction.Rollback();
             throw;
         }
+    }
+
+    public async Task<bool> ExistsAsync(int id)
+    {
+        using var connection = dbContext.GetConnection();
+        return await connection.ExecuteScalarAsync<bool>(
+            """
+            SELECT COUNT(*) > 0
+            FROM "Auth"."Users"
+            WHERE "Id" = @Id
+            """,
+            new { Id = id });
     }
 }
